Move exposure manager debugging to another file
