+++ ios/BT/ExposureManager.swift
@@ -59,13 +59,13 @@ final class ExposureManager: NSObject {
    that is done in the deinit.
   */
 
-  fileprivate let manager: ExposureNotificationManager
-  private let apiClient: APIClient
-  private let btSecureStorage: BTSecureStorage
-  private let bgTaskScheduler: BackgroundTaskScheduler
-  private let notificationCenter: NotificationCenter
-  private let userNotificationCenter: UserNotificationCenter
-  private let keychainService: KeychainService
+  public let manager: ExposureNotificationManager
+  public let apiClient: APIClient
+  public let btSecureStorage: BTSecureStorage
+  public let bgTaskScheduler: BackgroundTaskScheduler
+  public let notificationCenter: NotificationCenter
+  public let userNotificationCenter: UserNotificationCenter
+  public let keychainService: KeychainService
 
   init(exposureNotificationManager: ExposureNotificationManager = ENManager(),
        apiClient: APIClient = BTAPIClient.shared,
@@ -190,9 +190,8 @@ final class ExposureManager: NSObject {
   // MARK: == Diagnosis Keys ==
 
   typealias ExposureKeysDictionaryArray = [[String: Any]]
-  /// Requests the temporary exposure keys used by this device to share with a server. Returns an array of the exposures keys as dictionary or and error if the underlying
-  /// API fails
 
+  /// Requests the temporary exposure keys used by this device to share with a server. Returns an array of the exposures keys as dictionary or and error if the underlying API fails
   @objc func fetchExposureKeys(callback: @escaping (ExposureKeysDictionaryArray?, ExposureManagerError?) -> Void) {
     getDiagnosisKeys(transform: { (keys) -> ExposureKeysDictionaryArray in
       (keys ?? []).map { $0.asDictionary }
@@ -564,98 +563,10 @@ private extension ExposureManager {
     }
   }
 
-
   func cleanup() {
     // Delete downloaded files from file system
     localUncompressedURLs.cleanup()
     localUncompressedURLs = []
     downloadedPackages = []
-  }
-}
-
-protocol ExposureManagerDebuggable {
-  func handleDebugAction(_ action: DebugAction,
-                               resolve: @escaping RCTPromiseResolveBlock,
-                               reject: @escaping RCTPromiseRejectBlock)
-}
-
-extension ExposureManager: ExposureManagerDebuggable {
-
-  @objc func handleDebugAction(_ action: DebugAction,
-                               resolve: @escaping RCTPromiseResolveBlock,
-                               reject: @escaping RCTPromiseRejectBlock) {
-    switch action {
-    case .fetchDiagnosisKeys:
-      manager.getDiagnosisKeys { (keys, error) in
-        if let error = error {
-          reject(error.localizedDescription, "Failed to get exposure keys", error)
-        } else {
-          resolve(keys!.map { $0.asDictionary })
-        }
-      }
-    case .detectExposuresNow:
-      guard btSecureStorage.userState.remainingDailyFileProcessingCapacity > 0 else {
-        let hoursRemaining = 24 - Date.hourDifference(from: btSecureStorage.userState.dateLastPerformedFileCapacityReset ?? Date(),
-                                                      to: Date())
-        reject("Time window Error.",
-               "You have reached the exposure file submission limit. Please wait \(hoursRemaining) hours before detecting exposures again.",
-          GenericError.unknown)
-        return
-      }
-
-      detectExposures { result in
-        switch result {
-        case .success(let numberOfFilesProcessed):
-          resolve("Exposure detection successfully executed. Processed \(numberOfFilesProcessed) files.")
-        case .failure(let exposureError):
-          reject(exposureError.localizedDescription, exposureError.errorDescription, exposureError)
-        }
-      }
-    case .simulateExposureDetectionError:
-      btSecureStorage.exposureDetectionErrorLocalizedDescription = "Unable to connect to server."
-      postExposureDetectionErrorNotification("Simulated Error")
-      resolve(String.genericSuccess)
-    case .simulateExposure:
-      let exposure = Exposure(id: UUID().uuidString,
-                              date: Date().posixRepresentation - Int(TimeInterval.random(in: 0...13)) * 24 * 60 * 60 * 1000,
-                              duration: TimeInterval(1),
-                              totalRiskScore: .random(in: 1...8),
-                              transmissionRiskLevel: .random(in: 0...7))
-      btSecureStorage.storeExposures([exposure])
-      let content = UNMutableNotificationContent()
-      content.title = String.newExposureNotificationTitle.localized
-      content.body = String.newExposureNotificationBody.localized
-      content.sound = .default
-      let request = UNNotificationRequest(identifier: "identifier", content: content, trigger: nil)
-      userNotificationCenter.add(request) { error in
-        DispatchQueue.main.async {
-          if let error = error {
-            print("Error showing error user notification: \(error)")
-          }
-        }
-      }
-      resolve("Exposures: \(btSecureStorage.userState.exposures)")
-    case .fetchExposures:
-      resolve(currentExposures)
-    case .getAndPostDiagnosisKeys:
-      getAndPostDiagnosisKeys(certificate: .default, HMACKey: .default) { (success, error) in
-        if let error = error {
-          reject(error.errorCode, error.localizedMessage, error.underlyingError)
-        } else {
-          resolve(success)
-        }
-      }
-    case .resetExposures:
-      btSecureStorage.exposures = List<Exposure>()
-      resolve("Exposures: \(btSecureStorage.exposures.count)")
-    case .toggleENAuthorization:
-      let enabled = manager.exposureNotificationEnabled ? false : true
-      requestExposureNotificationAuthorization(enabled: enabled) { result in
-        resolve("EN Enabled: \(self.manager.exposureNotificationEnabled)")
-      }
-    case .showLastProcessedFilePath:
-      let path = btSecureStorage.userState.urlOfMostRecentlyDetectedKeyFile
-      resolve(path)
-    }
   }
 }
